fix(cli): make repl execute all commands in the edge vm instead of copying the values
This fixes an issue that came up in #62: when running the REPL and
comparing a value to `ArrayBuffer`, the globally-available ArrayBuffer
is actually the host one, and not the edge-runtime one.

This makes sure that every value is coming from the VM. Unfortunately,
it also means that we don't support top-level async/await, but I think
we can make it work somehow.